Refactor test file imports to remove unnecessary paths
